Simplify page (single/range) allocation.
The previous algorithm was:
- Grow file by requested size
- Write one byte to the end of the newly allocated range
- Flush

The current algorithm adds the need to store the total page count
withing the FileHeader. To allocate and claim new pages, this value is
then updated and the updated page is written. This approach is much more
similar to the handling of updates of other page types, because now the page
number claims are also WAL-ready.