Add location to in-memory pages.
A `noid::backend::page::Page` now contains its location (PageNumber) when
deserialized. The location of new pages is claimed by their respective
builders using the associated Pager. Claiming a page is the callers'
responsibility.

Having the location available when loaded into memory allows updated
versions of a Page to overwrite the location of their
predecessors if so desired.